Quarantine Chaos for Black Ops III
Inspired from QC from MW2, this is a multiplayer based zombie game mode.
One player is chosen as the alpha zombie and tries to turn the other
humans in to zombies while the humans fight for their lives.

I'll be making periodic updated to this repository once things get
tested and new features are added.